+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Automatic temporal extent calculation for collections. When serving collections via the `/collections` 
+  and `/collections/{collectionId}` endpoints, if a collection does not have a temporal extent defined, 
+  the server will automatically calculate it from the earliest and latest items in the collection. To use 
+  this feature, simply omit the `extent.temporal.interval` field when ingesting a collection.
 - Asset proxying for generating pre-signed S3 URLs through proxy endpoints `GET
   /collections/{collectionId}/items/{itemId}/assets/{assetKey}` and `GET
   /collections/{collectionId}/assets/{assetKey}`.